pyproject.toml
[tool.poetry]
name = "optimization1"
version = "0.1.0"
description = ""
authors = ["Sotnikov Arkadiy <cotnikoarkady@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.8.1"
hydra-core = "^1.1.2"
matplotlib = "^3.5.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
